+++ lib/cocoadex/docset_helper.rb
@@ -1,5 +1,7 @@
 
 module Cocoadex
+
+  # Helper for finding and indexing DocSets
   class DocSetHelper
 
     ROOT_PATHS = [
@@ -15,7 +17,9 @@ module Cocoadex
 
     def self.docset_paths
       @paths ||= begin
-        ROOT_PATHS.map { |path| Dir.glob(File.expand_path(path)+'/*/') }.flatten
+        ROOT_PATHS.map do |path|
+          Dir.glob(File.expand_path(path)+'/*/')
+        end.flatten
       end
     end
 
@@ -28,8 +32,8 @@ module Cocoadex
       end
 
       if docsets.size > 0
-        Keyword.write(:overwrite)
-        Keyword.generate_tags!
+        Tokenizer.persist
+        CompletionHelper.generate_tags!
         write(docsets)
       end
       logger.info "Done! #{docsets.size} DocSet#{docsets.size == 1 ? '':'s'} indexed."
@@ -45,7 +49,7 @@ module Cocoadex
 
     def self.write docsets
       @docsets = docsets
-      Serializer.write(data_path, docsets, :overwrite)
+      Serializer.write_array(data_path, docsets, :overwrite)
     end
   end
 end