+++ lib/cocoadex/models/callback.rb
@@ -1,0 +1,12 @@
+
+module Cocoadex
+  class Callback < Function
+    def type
+      "Callback"
+    end
+
+    def discussion
+      @discussion
+    end
+  end
+end