+++ lib/cocoadex/models/entity.rb
@@ -2,8 +2,10 @@ module Cocoadex
   # A top level element, roughly equivalent to one
   # page of documentation
   class Entity < Element
+    attr_reader :path
 
     def initialize path
+      @path    = path
       text     = clean(IO.read(path, :mode => 'rb'))
       document = Nokogiri::HTML(text)
       parse(document)
@@ -17,6 +19,10 @@ module Cocoadex
 
     def strip text
       text.gsub("&#xA0;&#xA0;","")
+    end
+
+    def section_by_title doc, title
+      doc.css("section").to_a.detect {|s| s.css("h2.jump").text == title }
     end
   end
 end