+++ lib/cocoadex/serializer.rb
@@ -1,5 +1,7 @@
 
 module Cocoadex
+
+  # Move data to and from disk
   class Serializer
     SEPARATOR = "--__--"
 
@@ -13,6 +15,7 @@ module Cocoadex
       array
     end
 
+    # Ensure parent paths exist
     def self.check_path path
       unless File.exists? File.dirname(path)
         FileUtils.mkdir_p File.dirname(path)
@@ -20,29 +23,36 @@ module Cocoadex
     end
 
     # Write text to a file
-    def self.write_text path, text
+    def self.write_text path, text, style=:overwrite
       check_path path
+
+      mode = style_to_mode(style)
+
       File.open(path, 'w') do |file|
         file.print text
       end
     end
 
     # Write a cache Array as a serialized file
-    def self.write path, array, style
+    def self.write_array path, array, style
       check_path path
 
-      mode = case style
-        when :append then 'a'
-        when :overwrite then 'w'
-        else
-          raise "Unknown file mode: #{style}"
-      end
+      mode = style_to_mode(style)
 
       File.open(path, mode) do |file|
         array.each do |object|
           file.print(Marshal.dump(object))
           file.print SEPARATOR
         end
+      end
+    end
+
+    def self.style_to_mode style
+      case style
+        when :append then 'a'
+        when :overwrite then 'w'
+        else
+          raise "Unknown file mode: #{style}"
       end
     end
   end